requirements.txt
# 项目依赖

# 深度学习框架
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0

# 计算机视觉
timm>=0.6.0  # Vision Transformer和其他预训练模型
albumentations>=1.3.0  # 图像增强
opencv-python>=4.5.0
Pillow>=9.0.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# 医学图像处理
nibabel>=3.2.0  # NIfTI格式MRI图像处理
SimpleITK>=2.1.0  # 医学图像处理
monai>=0.9.0  # 医学图像深度学习框架

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
tensorboard>=2.8.0
wandb>=0.12.0  # 实验跟踪

# 工具
jupyter>=1.0.0
jupyterlab>=3.0.0
tqdm>=4.62.0
pyyaml>=6.0
h5py>=3.6.0

# 模型解释
grad-cam>=1.4.0  # 热力图可视化
captum>=0.5.0  # PyTorch模型可解释性
